code reorg in prep for working multidim IPF/QIS
