+++ src/IPF.cpp
@@ -145,3 +145,122 @@ bool IPF::computeErrors(std::vector<std::vector<double>>& diffs)
   return m_maxError < m_tol;
 }
 
+namespace wip {
+  
+// TODO perhaps seed should be an arg to solve instead of being passed in here
+IPF::IPF(/*const NDArray<double>& seed,*/ const index_list_t& indices, marginal_list_t& marginals)
+  : /*m_seed(std::move(seed)),*/ Microsynthesis(indices, marginals)
+{
+  // TODO check seed dims match those computed by base
+}
+
+NDArray<double>& IPF::solve()
+{
+  Index index_main(m_array.sizes());
+  
+  std::vector<MappedIndex> mappings;
+  mappings.reserve(m_marginals.size());
+  for (size_t k = 0; k < m_marginals.size(); ++k)
+  {
+    mappings.push_back(MappedIndex(index_main, m_indices[k]));
+  }
+
+  m_array.assign(1.0);
+
+  marginal_list_t diffs(m_marginals.size());
+  m_errors.resize(m_marginals.size());
+  
+  for (size_t k = 0; k < diffs.size(); ++k)
+  {
+    diffs[k].resize(m_marginals[k].sizes());
+    m_errors[k].resize(m_marginals[k].sizes());
+  }
+
+  m_conv = false;
+  for (m_iters = 0; !m_conv && m_iters < 3; ++m_iters)
+  {
+    rScale(/*m_array, m_marginals*/);
+    rDiff(diffs, m_array, m_marginals);
+
+    m_conv = computeErrors(diffs);
+  }
+
+  return m_array;
+}
+
+bool IPF::computeErrors(std::vector<NDArray<double>>& diffs)
+{
+  m_maxError = -std::numeric_limits<double>::max();
+
+  // // create mapped indices
+  // const std::vector<MappedIndex>& mapped = makeMappings(main_index);
+  for (size_t k = 0; k < diffs.size(); ++k)
+  {
+    for (Index index(diffs[k].sizes()); !index.end(); ++index)
+    {
+      double e = std::fabs(diffs[k][index]);
+      m_errors[k][index] = e;
+      m_maxError = std::max(m_maxError, e);
+    }
+  }
+
+  return m_maxError < m_tol;
+}
+
+void IPF::rScale()
+{
+  // for (size_t d = 0; d < result.dim(); ++d)
+  // {
+  //   const std::vector<double>& r = reduce<double>(result, d);
+  //   for (size_t p = 0; p < marginals[d].size(); ++p)
+  //   {
+  //     for (Index index(result.sizes(), { d, p }); !index.end(); ++index)
+  //     {
+  //       const std::vector<int64_t>& ref = index;
+  //       // avoid division by zero (assume 0/0 -> 0)
+  //       if (r[p] == 0.0 && marginals[d][ref[d]] != 0.0)
+  //         throw std::runtime_error("div0 in rScale with m>0");
+  //       if (r[p] != 0.0)
+  //         result[index] *= marginals[d][ref[d]] / r[p];
+  //       else
+  //         result[index] = 0.0;
+  //     }
+  //   }
+  // }
+  for (size_t k = 0; k < m_indices.size(); ++k)
+  {
+    const NDArray<double>& r = reduce<double>(m_array, m_indices[k]);
+    std::cout << k << ":";
+    print(r.rawData(), r.storageSize());
+
+
+    Index main_index(m_array.sizes());
+    //std::cout << m_array.sizes()[m_indices[1-k][0]] << std::endl;
+    for (MappedIndex oindex(main_index, invert(m_array.dim(), m_indices[k])); !oindex.end(); ++oindex)
+    {
+      for (MappedIndex index(main_index, m_indices[k]); !index.end(); ++index)
+      {
+        print((std::vector<int64_t>)main_index);
+        if (r[index] == 0.0 && m_marginals[k][index] != 0.0)
+          throw std::runtime_error("div0 in rScale with m>0");
+        if (r[index] != 0.0)
+          m_array[main_index] *= m_marginals[k][index] / r[index];
+        else
+          m_array[main_index] = 0.0;
+      }
+    }
+    // reset the main index
+    //main_index.reset();
+  }
+}
+
+void IPF::rDiff(std::vector<NDArray<double>>& diffs, const NDArray<double>& result, const std::vector<NDArray<double>>& marginals)
+{
+  int64_t n = m_indices.size();
+  for (int64_t k = 0; k < n; ++k)
+    diff(reduce<double>(result, m_indices[k]), marginals[k], diffs[k]);
+}
+
+
+} // wip
+
